refactor: 컨트롤러 함수 분리
